+++ Carbon.IIS/Carbon.IIS.psm1
@@ -92,12 +92,21 @@ if( -not (Test-MSWebAdministrationLoaded) )
 }
 
 $sm = [Microsoft.Web.Administration.ServerManager]::New()
-if( -not $sm -or $null -eq $sm.Sites )
+try
 {
-    Write-Error -Message "Carbon.IIS is not supported on this version of PowerShell." -ErrorAction Stop
-    return
+    if( -not $sm -or $null -eq $sm.Sites )
+    {
+        Write-Error -Message "Carbon.IIS is not supported on this version of PowerShell." -ErrorAction Stop
+        return
+    }
 }
-
+finally
+{
+    if( $sm )
+    {
+        $sm.Dispose()
+    }
+}
 # We successfully loaded Microsoft.Web.Administration assembly, so remove the errors we encountered trying to do so.
 for( $idx = $Global:Error.Count ; $idx -gt $numErrorsAtStart ; --$idx )
 {