+++ Carbon.IIS/Functions/Get-CIisApplication.ps1
@@ -14,8 +14,6 @@ function Get-CIisApplication
      * `CommitChanges()` - Persists any configuration changes made to the object back into IIS's configuration files.
      * `PhysicalPath { get; }` - The physical path to the application.
 
-    Beginning with Carbon 2.0.1, this function is available only if IIS is installed.
-
     .OUTPUTS
     Microsoft.Web.Administration.Application.
 
@@ -37,36 +35,34 @@ function Get-CIisApplication
     [CmdletBinding()]
     [OutputType([Microsoft.Web.Administration.Application])]
     param(
-        [Parameter(Mandatory=$true)]
-        [string]
         # The site where the application is running.
-        $SiteName,
-        
-        [Parameter()]
-        [Alias('Name')]
-        [string]
-        # The name of the application.  Default is to return all applications running under the website `$SiteName`.
-        $VirtualPath
+        [Parameter(Mandatory)]
+        [String] $SiteName,
+
+        # The path/name of the application. Default is to return all applications running under the website given by
+        # the `SiteName` parameter.
+        [String] $VirtualPath
     )
 
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $site = Get-CIisWebsite -SiteName $SiteName
+    $site = Get-CIisWebsite -Name $SiteName
     if( -not $site )
     {
         return
     }
 
+    $VirtualPath = $VirtualPath | ConvertTo-CIisVirtualPath
+
     $site.Applications |
         Where-Object {
-            if( $VirtualPath )
+            if( $PSBoundParameters.ContainsKey('VirtualPath') )
             {
-                return ($_.Path -eq "/$VirtualPath")
+                return ($_.Path -eq $VirtualPath)
             }
             return $true
-        } | 
+        } |
         Add-IisServerManagerMember -ServerManager $site.ServerManager -PassThru
 }