+++ Carbon.IIS/Functions/Install-CIisWebsite.ps1
@@ -163,7 +163,7 @@ function Install-CIisWebsite
     {
         Write-Verbose -Message ('Creating website ''{0}'' ({1}).' -f $Name,$PhysicalPath)
         $firstBinding = $Binding | Select-Object -First 1 | ConvertTo-Binding
-        $mgr = New-Object 'Microsoft.Web.Administration.ServerManager'
+        $mgr = New-CIisServerManager
         $site = $mgr.Sites.Add( $Name, $firstBinding.Protocol, $firstBinding.BindingInformation, $PhysicalPath )
         $mgr.CommitChanges()
     }
@@ -246,7 +246,7 @@ function Install-CIisWebsite
     $website = $null
     do
     {
-        $website = Get-CIisWebsite -SiteName $Name
+        $website = Get-CIisWebsite -Name $Name
         $tries += 1
         if($website.State -ne 'Unknown')
         {